dung-refactor errors related to project structure
